ExploreParks updates dynamically based on input
Updates the ExplorePark component to update page data based off
this.state.currentPark. setState on the api call chain sets the
currentPark and image keys to the first park stored in the parks key.
The jsx <select> element contains the names of all the parks in
this.state.parks and the index of those parks in it's options.
When the value of <select> changes, the handleChange method sets the
state of the image and currentPark key to the selected input's index
within this.state.parks. As a result, currentPark stores the data of
this.state.parks[<parkIndex>]