Adds Logic for favorite button
If a user who is not signed in clicks on "Add to Favorite Parks"
button, they are redirected to Sign In.
Will have to add a sign up button on SignIn component.
Still need to implement logic if user is signed in, and
change the button to "remove from" is park is already on their userList