Adds logic and fetch for toggling a favoritePark
Creates the api call and event chain for adding or removing a parkCode
from a user's favoriteParks list (userList). The event handler
addFavorite leverages a ternary to differentiate a signed-in user from
an anonymous (not-signed-in) one. The anonymous user is asked to sign in
while the authenticated user will run addToParks with the user and
picked park as arguments. Another ternary in parksApi determines whether
or not the user has a userList, and either calls the
favoriteParks/create end point, or the /updateOne endpoint with the user
and park passed to the api. Upon a successful create or updateOne,
setUser is called so that a local copy of the userList is stored.