Creates skeleton structure for favorites list
Creates the Link, Route and Favorites Component so that a user can see a
list of their favoriteParks. Breaks down the list into
ParkItem components. Currently uses the parks state, which contains all
parks returned from the api. This will be revised after an api refactor
so that only parks on the user's favoriteParks list are rendered.