Adds a message for user's who have an empty list
As opposed to first time users, this feature is for user's who started
a favoriteParks list, but then removed all parks from the list.
Along with a change to the api, which returns res.data === null in this
case, the user will now see a message that prompts them to add some
parks to their favoriteParks List.